wind direction func in a separate module
